+++ src/base/RenderObject.h
@@ -40,18 +40,18 @@ namespace OpenGC {
     RenderObject();
     virtual ~RenderObject();
 		
-    virtual void SetUnitsPerPixel(const double unitsPerPixel);
-    virtual void SetScale(const double xScale, const double yScale);
-    void SetPosition(const double xPos, const double yPos);
+    virtual void SetUnitsPerPixel(const float unitsPerPixel);
+    virtual void SetScale(const float xScale, const float yScale);
+    void SetPosition(const float xPos, const float yPos);
     void SetParentRenderObject(RenderObject* pObject);
-    std::pair<double, double> GetPhysicalPosition(void);
+    std::pair<float, float> GetPhysicalPosition(void);
 
     /** Called by framework when a mouse click occurs (x/y in pixel coordinates) */
     void HandleMouseButton(const int button, const int state, const int x, const int y);
     /** Called when a mouse "down" event occurs (x/y in physical coordinates) */
-    virtual void OnMouseDown(const int button, const double physicalX, const double physicalY);
+    virtual void OnMouseDown(const int button, const float physicalX, const float physicalY);
     /** Called when a mouse "up" event occurs (x/y in physical coordinates) */
-    virtual void OnMouseUp(const int button, const double physicalX, const double physicalY);	
+    virtual void OnMouseUp(const int button, const float physicalX, const float physicalY);	
     /** Called when a key is pressed on the keyboard and OpenGC has focus */
     virtual void OnKeyboard(const int keycode, const int modifiers);
 
@@ -61,17 +61,17 @@ namespace OpenGC {
 
   protected:
     /** Set by the render window to describe pixel-realspace conversions */
-    double m_UnitsPerPixel;
+    float m_UnitsPerPixel;
     /** Parent object, for cascading position information */
     RenderObject *m_pParentRenderObject;
-    /** 1.0=normal scale, 0.5=half, 2=double, etc. */
-    std::pair<double, double> m_Scale;
+    /** 1.0=normal scale, 0.5=half, 2=float, etc. */
+    std::pair<float, float> m_Scale;
     /** Position in mm in the render window */
-    std::pair<double, double> m_PhysicalPosition;
+    std::pair<float, float> m_PhysicalPosition;
     /** Placement in render window in pixel units */
     std::pair<unsigned int, unsigned int> m_PixelPosition;
     /** Size in mm, defined by derived classes, NOT initialized */
-    std::pair<double, double> m_PhysicalSize;
+    std::pair<float, float> m_PhysicalSize;
     /** Size in render window in pixel units */
     std::pair<unsigned int, unsigned int> m_PixelSize;
   };
